// src/types/Card.ts
export interface Card {
  cardId: string;
  oracleId: string;
  mtgoId?: number;
  mtgoFoilId?: number;
  tcgplayerId?: number;
  cardmarketId?: number;
  name: string;
  lang: string;
  releasedAt: string;
  uri: string;
  scryfallUri: string;
  layout: string;
  highresImage: boolean;
  imageStatus: string;
  smallImageUri: string;
  normalImageUri: string;
  largeImageUri: string;
  pngImageUri: string;
  artCropImageUri: string;
  borderCropImageUri: string;
  manaCost: string;
  cmc: number;
  typeLine: string;
  oracleText: string;
  colors: string[];
  colorIdentity: string[];
  keywords: string[];
  games: string[];
  finishes: string[];
  reserved: boolean;
  foil: boolean;
  nonfoil: boolean;
  oversized: boolean;
  promo: boolean;
  reprint: boolean;
  variation: boolean;
  setId: string;
  setName: string;
  setType: string;
  setUri: string;
  setSearchUri: string;
  scryfallSetUri: string;
  rulingsUri: string;
  printsSearchUri: string;
  collectorNumber: string;
  digital: boolean;
  rarity: string;
  flavorText?: string;
  cardBackId: string;
  artist: string;
  artistIds: string[];
  illustrationId: string;
  borderColor: string;
  frame: string;
  fullArt: boolean;
  textless: boolean;
  booster: boolean;
  storySpotlight: boolean;
  edhrecRank?: number;
  usdPrice?: string;
  usdFoilPrice?: string;
  usdEtchedPrice?: string | null;
  eurPrice?: string;
  eurFoilPrice?: string;
  tixPrice?: string;
  tcgplayerInfiniteArticlesUri: string;
  tcgplayerInfiniteDecksUri: string;
  edhrecUri: string;
  tcgplayerUri: string;
  cardmarketUri: string;
  cardhoarderUri: string;
  standardLegal: boolean;
  futureLegal: boolean;
  historicLegal: boolean;
  timelessLegal: boolean;
  gladiatorLegal: boolean;
  pioneerLegal: boolean;
  explorerLegal: boolean;
  modernLegal: boolean;
  legacyLegal: boolean;
  pauperLegal: boolean;
  vintageLegal: boolean;
  pennyLegal: boolean;
  commanderLegal: boolean;
  oathbreakerLegal: boolean;
  brawlLegal: boolean;
  alchemyLegal: boolean;
  paupercommanderLegal: boolean;
  duelLegal: boolean;
  oldschoolLegal: boolean;
  premodernLegal: boolean;
  predhLegal: boolean;
}

export interface CardResponse {
  content: Card[];
  pageable: {
    pageNumber: number;
    pageSize: number;
  };
  totalElements: number;
  totalPages: number;
  numberOfElements: number;
  empty: boolean;
}